create analytic grouping module and style 100%
